making authorized requests

+++ frontend/src/components/WorkoutDetails.js
@@ -1,12 +1,18 @@
 import { useWorkoutContext } from "../hooks/useWorkoutContext";
 import { formatDistanceToNow } from "date-fns";
-
+import { useAuthContext } from "../hooks/useAuthContext";
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutContext();
-
+  const { user } = useAuthContext();
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        " Authorization": `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
 
@@ -23,7 +29,9 @@ const WorkoutDetails = ({ workout }) => {
       <p>
         <strong>Reps: {workout.reps}</strong>
       </p>
-      <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
+      <p>
+        {formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}
+      </p>
       <span className="material-symbols-outlined" onClick={handleClick}>
         delete
       </span>
